requirements.txt
fastapi~=0.128.0
uvicorn
requests
uuid~=1.30
pydantic~=2.12.5
qiskit==1.4.3
qiskit_aer==0.17.2
qiskit_qasm3_import
python-dotenv